+++ TrafficDataContracts/src/main/java/fi/ahto/example/traffic/data/contracts/internal/VehicleActivity.java
@@ -204,19 +204,19 @@ public class VehicleActivity implements Serializable {
         this.source = source;
     }
 
-    public Double getSpeed() {
+    public Float getSpeed() {
         return speed;
     }
 
-    public void setSpeed(Double speed) {
+    public void setSpeed(Float speed) {
         this.speed = speed;
     }
 
-    public Double getBearing() {
+    public Float getBearing() {
         return bearing;
     }
 
-    public void setBearing(Double bearing) {
+    public void setBearing(Float bearing) {
         this.bearing = bearing;
     }
 
@@ -316,19 +316,19 @@ public class VehicleActivity implements Serializable {
         this.direction = direction;
     }
 
-    public Double getLongitude() {
+    public Float getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Double longitude) {
+    public void setLongitude(Float longitude) {
         this.longitude = longitude;
     }
 
-    public Double getLatitude() {
+    public Float getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Double latitude) {
+    public void setLatitude(Float latitude) {
         this.latitude = latitude;
     }
 
@@ -347,9 +347,9 @@ public class VehicleActivity implements Serializable {
     @JsonProperty("Direction")
     private String direction;
     @JsonProperty("Longitude")
-    private Double longitude;
+    private Float longitude;
     @JsonProperty("Latitude")
-    private Double latitude;
+    private Float latitude;
     @JsonProperty("Delay")
     private Integer delay;
     @JsonProperty("InternalLineId")
@@ -373,9 +373,9 @@ public class VehicleActivity implements Serializable {
     @JsonProperty("LineHasChanged")
     private boolean lineHasChanged;
     @JsonProperty("Bearing")
-    private Double bearing;
+    private Float bearing;
     @JsonProperty("Speed")
-    private Double speed;
+    private Float speed;
     @JsonProperty("Source")
     private String source;
     @JsonProperty("NextStopId")
